OtherProjects/rag-knowledge-assistant/requirement.txt
langchain 
openai 
faiss-cpu 
sentence-transformers 
pypdf
llama-index
streamlit
langchain-google-genai
langchain-community
